requirements.txt
Flask
werkzeug
Pillow
deepface
transformers
spotipy
